# extension/query/api/assets/swagger/v1alpha1/query_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "v1alpha1/query_service.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/apis/logs/v1alpha1/logging": {
      "get": {
        "summary": "SearchTraces searches for traces.\nSee GetTrace for JSON unmarshalling.",
        "operationId": "QueryService_SearchLogs",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LogsData"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "QueryService"
        ]
      }
    },
    "/apis/traces/v1alpha1/operations": {
      "get": {
        "summary": "GetOperations returns operation names.",
        "operationId": "QueryService_GetOperations",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1alpha1GetOperationsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "service",
            "description": "Required service name.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "spanKind",
            "description": "Optional span kind.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "QueryService"
        ]
      }
    },
    "/apis/traces/v1alpha1/services": {
      "get": {
        "summary": "GetServices returns service names.",
        "operationId": "QueryService_GetServices",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1alpha1ResourcesData"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "QueryService"
        ]
      }
    },
    "/apis/traces/v1alpha1/trace": {
      "get": {
        "summary": "SearchTraces searches for traces.\nSee GetTrace for JSON unmarshalling.",
        "operationId": "QueryService_SearchTraces",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TracesData"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "query.serviceName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "query.operationName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "query.startTime",
            "description": "Span min start time in. REST API uses RFC-3339ns format. Required.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "query.endTime",
            "description": "Span max start time. REST API uses RFC-3339ns format. Required.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "query.durationMin",
            "description": "Span min duration. REST API uses Golang's time format e.g. 10s.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "query.durationMax",
            "description": "Span max duration. REST API uses Golang's time format e.g. 10s.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "query.numTraces",
            "description": "Maximum number of traces in the response.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "QueryService"
        ]
      }
    },
    "/apis/traces/v1alpha1/trace/{traceId}": {
      "get": {
        "summary": "GetTrace returns a single trace.\nNote that the JSON response over HTTP is wrapped into result envelope \"{\"result\": ...}\"\nIt means that the JSON response cannot be directly unmarshalled using JSONPb.\nThis can be fixed by first parsing into user-defined envelope with standard JSON library\nor string manipulation to remove the envelope. Alternatively generate objects using OpenAPI.",
        "operationId": "QueryService_GetTrace",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TracesData"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "traceId",
            "description": "Hex encoded 64 or 128 bit trace ID.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "QueryService"
        ]
      }
    }
  },
  "definitions": {
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "v1KeyValue": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "$ref": "#/definitions/v1AnyValue"
        }
      },
      "description": "KeyValue is a key-value pair that is used to store Span attributes, Link\nattributes, etc."
    },
...
    "v1LogsData": {
      "type": "object",
      "properties": {
        "resourceLogs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ResourceLogs"
          },
          "description": "An array of ResourceLogs.\nFor data coming from a single resource this array will typically contain\none element. Intermediary nodes that receive data from multiple origins\ntypically batch the data before forwarding further and in that case this\narray will contain multiple elements."
        }
      },
      "description": "LogsData represents the logs data that can be stored in a persistent storage,\nOR can be embedded by other protocols that transfer OTLP logs data but do not\nimplement the OTLP protocol.\n\nThe main difference between this message and collector protocol is that\nin this message there will not be any \"control\" or \"metadata\" specific to\nOTLP protocol.\n\nWhen new fields are added into this message, the OTLP request MUST be updated\nas well."
    },
    "v1Resource": {
      "type": "object",
      "properties": {
        "attributes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1KeyValue"
          },
          "description": "Set of attributes that describe the resource.\nAttribute keys MUST be unique (it is not allowed to have more than one\nattribute with the same key)."
        },
        "droppedAttributesCount": {
          "type": "integer",
          "format": "int64",
          "description": "dropped_attributes_count is the number of dropped attributes. If the value is 0, then\nno attributes were dropped."
        }
      },
      "description": "Resource information."
    },
    "v1ResourceLogs": {
      "type": "object",
      "properties": {
        "resource": {
          "$ref": "#/definitions/v1Resource",
          "description": "The resource for the logs in this message.\nIf this field is not set then resource info is unknown."
        },
        "scopeLogs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ScopeLogs"
          },
          "description": "A list of ScopeLogs that originate from a resource."
        },
        "schemaUrl": {
          "type": "string",
          "description": "This schema_url applies to the data in the \"resource\" field. It does not apply\nto the data in the \"scope_logs\" field which have their own schema_url field."
        }
      },
      "description": "A collection of ScopeLogs from a Resource."
    },
    "v1ResourceSpans": {
      "type": "object",
      "properties": {
        "resource": {
          "$ref": "#/definitions/v1Resource",
          "description": "The resource for the spans in this message.\nIf this field is not set then no resource info is known."
        },
        "scopeSpans": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ScopeSpans"
          },
          "description": "A list of ScopeSpans that originate from a resource."
        },
        "schemaUrl": {
          "type": "string",
          "description": "This schema_url applies to the data in the \"resource\" field. It does not apply\nto the data in the \"scope_spans\" field which have their own schema_url field."
        }
      },
      "description": "A collection of ScopeSpans from a Resource."
    },
    "v1ScopeLogs": {
      "type": "object",
      "properties": {
        "scope": {
          "$ref": "#/definitions/v1InstrumentationScope",
          "description": "The instrumentation scope information for the logs in this message.\nSemantically when InstrumentationScope isn't set, it is equivalent with\nan empty instrumentation scope name (unknown)."
        },
        "logRecords": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1LogRecord"
          },
          "description": "A list of log records."
        },
        "schemaUrl": {
          "type": "string",
          "description": "This schema_url applies to all logs in the \"logs\" field."
        }
      },
      "description": "A collection of Logs produced by a Scope."
    },
    "v1ScopeSpans": {
      "type": "object",
      "properties": {
        "scope": {
          "$ref": "#/definitions/v1InstrumentationScope",
          "description": "The instrumentation scope information for the spans in this message.\nSemantically when InstrumentationScope isn't set, it is equivalent with\nan empty instrumentation scope name (unknown)."
        },
        "spans": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Span"
          },
          "description": "A list of Spans that originate from an instrumentation scope."
        },
        "schemaUrl": {
          "type": "string",
          "description": "This schema_url applies to all spans and span events in the \"spans\" field."
        }
      },
      "description": "A collection of Spans produced by an InstrumentationScope."
    },
...
    "v1TracesData": {
      "type": "object",
      "properties": {
        "resourceSpans": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ResourceSpans"
          },
          "description": "An array of ResourceSpans.\nFor data coming from a single resource this array will typically contain\none element. Intermediary nodes that receive data from multiple origins\ntypically batch the data before forwarding further and in that case this\narray will contain multiple elements."
        }
      },
      "description": "TracesData represents the traces data that can be stored in a persistent storage,\nOR can be embedded by other protocols that transfer OTLP traces data but do\nnot implement the OTLP protocol.\n\nThe main difference between this message and collector protocol is that\nin this message there will not be any \"control\" or \"metadata\" specific to\nOTLP protocol.\n\nWhen new fields are added into this message, the OTLP request MUST be updated\nas well."
    },
    "v1alpha1GetOperationsResponse": {
      "type": "object",
      "properties": {
        "names": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "description": "Response object to get operation names."
    },
    "v1alpha1ResourcesData": {
      "type": "object",
      "properties": {
        "resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Resource"
          },
          "description": "An array of ResourceMetrics.\nFor data coming from a single resource this array will typically contain\none element. Intermediary nodes that receive data from multiple origins\ntypically batch the data before forwarding further and in that case this\narray will contain multiple elements."
        }
      }
    },
...
  }
}